week-2/requirements.txt
python-dotenv
psycopg2-binary
confluent_kafka